feat: 수익률을 출력하는 함수 구현

+++ src/main/java/lotto/domain/ScoreInfo.java
@@ -1,21 +1,21 @@
 package lotto.domain;
 
 public enum ScoreInfo {
-    THREE(5, 3, 0, 5000, "3개 일치 (5,000원) - "),
-    FOUR(4, 4, 0, 50000, "4개 일치 (50,000원) - "),
-    FIVE(3, 5, 0, 1500000, "5개 일치 (1,500,000원) - "),
-    FIVE_ALPHA(2, 6, 1, 30000000, "5개 일치, 보너스 볼 일치 (30,000,000원) - "),
-    SIX(1, 6, 0, 2000000000, "6개 일치 (2,000,000,000원) - "),
-    ELSE(0, 0, 0, 0, "");
+    THREE(5, 3, 0, 5000L, "3개 일치 (5,000원) - "),
+    FOUR(4, 4, 0, 50000L, "4개 일치 (50,000원) - "),
+    FIVE(3, 5, 0, 1500000L, "5개 일치 (1,500,000원) - "),
+    FIVE_ALPHA(2, 6, 1, 30000000L, "5개 일치, 보너스 볼 일치 (30,000,000원) - "),
+    SIX(1, 6, 0, 2000000000L, "6개 일치 (2,000,000,000원) - "),
+    ELSE(0, 0, 0, 0L, "");
 
 
     private final Integer rank;
     private final Integer correctCount;
     private final Integer correctBonusCount;
-    private final Integer winnerMoney;
+    private final Long winnerMoney;
     private final String message;
 
-    ScoreInfo(Integer rank, Integer correctCount, Integer correctBonusCount, Integer winnerMoney, String message) {
+    ScoreInfo(Integer rank, Integer correctCount, Integer correctBonusCount, Long winnerMoney, String message) {
         this.rank = rank;
         this.correctCount = correctCount;
         this.correctBonusCount = correctBonusCount;
@@ -27,7 +27,7 @@ public enum ScoreInfo {
         return rank;
     }
 
-    public Integer getWinnerMoney() {
+    public Long getWinnerMoney() {
         return winnerMoney;
     }
